disabled background tracking. Renaming of location logic
