+++ screens/DownloadMapScreen.tsx
@@ -4,7 +4,7 @@ import { StyleSheet, LayoutRectangle, Alert, View, Text, Modal, Button } from "r
 import { connect, ConnectedProps } from "react-redux";
 import { RootStackParamList } from "../shared/TypeDefinitions";
 import { createTrip } from "../shared/ActionCreators";
-import { startRouteTracking } from "../services/BackgroundLocationTracking";
+import { startRouteTracking } from "../services/LocationTracking";
 import MapView, { Region, UrlTile } from "react-native-maps";
 import { createMapDownloadTask, estimateDownloadTilesSize, IMapDownloadTask, ListenerSubscription } from "../services/MapDownload";
 import { FloatingAction } from "react-native-floating-action";
